refactor: apply PR review changes to rating logic and response format
