--- src/utils/AvgRating.js
@@ -1,37 +1,0 @@
-import prisma from '../config/DatabaseConfig.js';
-
-class AvgRatingHelper {
-  calculate = async (dest) => {
-    try {
-      const ratings = await prisma.rating.findMany({
-        where: {
-          dest_id: dest.dest_id,
-        },
-      });
-      const totalRating = ratings.length;
-      const sumRating = ratings.reduce((sum, rating) => sum + rating.rating, 0);
-      const avgRating = totalRating ? sumRating / totalRating : 0;
-      const roundedRating = parseFloat(avgRating.toFixed(1));
-      return {
-        id: dest.dest_id,
-        name: dest.name_dest,
-        description: dest.description,
-        img: dest.img,
-        location: dest.location,
-        avgRating: roundedRating,
-      };
-    } catch (error) {
-      console.error(error.message);
-      return {
-        id: dest.dest_id,
-        name: dest.name_dest,
-        description: dest.description,
-        img: dest.img,
-        location: dest.location,
-        avgRating: 0,
-      };
-    }
-  };
-}
-
-export default new AvgRatingHelper();
